Implemented:
- Hiding of /Data from sync clients.
- WebDAV access to folders/files shared via the web interface.

These last ones can be accessed via

https://data.deic.dk/remote.php/mydav/my_share

- assuming one has shared a folder/file under the name my_share.

This folder/file could have a different name and one could have a
folder/file by the share name in one's homedir. In that case, the above
url will give access to the folder/file with the name my_share.

To be sure which files one get, the mod_rewrite shortcuts defined in
.htaccess can be used:

https://data.deic.dk/files/my_share

and

https://data.deic.dk/public/my_share

The first one will give access to one's personal files, the second one
to the share.